Add debug logging for CreateTask params

+++ internal/database/schema.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS video_tasks (
     result_size BIGINT NOT NULL DEFAULT 0,
     manifest_path TEXT,
     
-    segment_sec DECIMAL(10,4) NOT NULL DEFAULT 8,
+    segment_sec DOUBLE PRECISION NOT NULL DEFAULT 8,
     frame_interval_sec INTEGER NOT NULL DEFAULT 2,
     frame_format VARCHAR(10) NOT NULL DEFAULT 'jpg',
     zip_format VARCHAR(10) NOT NULL DEFAULT 'zip',
